implement filter.limit in /context
This seems to be completely redundant with the 'limit' body parameter,
with the only difference being:

> The filter may be applied before or/and after the limit parameter -
filter> whichever the homeserver prefers.

This sentence seems to apply to the 'limit' body parameter, but not to
the 'limit' field on the filter. This was probably unintentional on the
part of the spec authors, but I've switched to using same 'load_limit'
pattern we're using elsewhere anyway.